# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customer/bookings": {
      "get": {
        "operationId": "BookingController_customerList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "BookingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/customer/bookings/{id}": {
      "get": {
        "operationId": "BookingController_customerOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings": {
      "get": {
        "operationId": "BookingController_opsList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "BookingController_opsCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpsBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/timeline": {
      "get": {
        "operationId": "BookingController_opsTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/revisions": {
      "get": {
        "operationId": "BookingController_opsRevisions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}": {
      "get": {
        "operationId": "BookingController_opsOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      },
      "patch": {
        "operationId": "BookingController_opsUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingController_opsCancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/reject": {
      "post": {
        "operationId": "BookingController_opsReject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/dashboard": {
      "get": {
        "operationId": "BookingController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/confirm": {
      "post": {
        "operationId": "BookingController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/assign": {
      "post": {
        "operationId": "BookingController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/bookings/{id}/advance": {
      "post": {
        "operationId": "BookingController_advance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/ops/calendar/week": {
      "get": {
        "operationId": "OperationsController_week",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/calendar/day": {
      "get": {
        "operationId": "OperationsController_day",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/me": {
      "get": {
        "operationId": "OperationsController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/vehicles": {
      "get": {
        "operationId": "OperationsController_vehicles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/customers": {
      "get": {
        "operationId": "OperationsController_customers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/customers/{id}/sites": {
      "get": {
        "operationId": "OperationsController_customerSites",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/ops/users": {
      "get": {
        "operationId": "OperationsController_users",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/v1/owner/audit": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/owner/audit/{id}": {
      "get": {
        "operationId": "AuditController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "audit"
        ]
      }
    }
  },
  "info": {
    "title": "Foresee Reservation API",
    "description": "Booking and operations API for Foresee Corporation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "serviceCode": {
            "type": "string",
            "example": "GREASE_TRAP"
          },
          "customerSiteId": {
            "type": "string",
            "format": "uuid"
          },
          "requestedDate": {
            "type": "string",
            "example": "2026-09-01"
          },
          "requestedStart": {
            "type": "string",
            "example": "09:00"
          },
          "requestedEnd": {
            "type": "string",
            "example": "11:30"
          },
          "estimatedVolume": {
            "type": "number",
            "example": 4
          },
          "volumeUnit": {
            "type": "string",
            "example": "ลบ.ม."
          },
          "customerNote": {
            "type": "string",
            "example": "เข้าทางประตูฝั่งตะวันออก"
          }
        },
        "required": [
          "serviceCode",
          "customerSiteId",
          "requestedDate",
          "requestedStart",
          "requestedEnd"
        ]
      },
      "CreateOpsBookingDto": {
        "type": "object",
        "properties": {
          "serviceCode": {
            "type": "string",
            "example": "GREASE_TRAP"
          },
          "customerSiteId": {
            "type": "string",
            "format": "uuid"
          },
          "requestedDate": {
            "type": "string",
            "example": "2026-09-01"
          },
          "requestedStart": {
            "type": "string",
            "example": "09:00"
          },
          "requestedEnd": {
            "type": "string",
            "example": "11:30"
          },
          "estimatedVolume": {
            "type": "number",
            "example": 4
          },
          "volumeUnit": {
            "type": "string",
            "example": "ลบ.ม."
          },
          "customerNote": {
            "type": "string",
            "example": "เข้าทางประตูฝั่งตะวันออก"
          },
          "customerOrganizationId": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "example": "ADMIN_PHONE",
            "enum": [
              "ADMIN_PHONE",
              "ADMIN_MANUAL"
            ]
          },
          "internalNote": {
            "type": "string",
            "example": "ลูกค้าโทรยืนยันกับคุณเอ"
          },
          "responsibleUserId": {
            "type": "string",
            "format": "uuid"
          },
          "contactName": {
            "type": "string",
            "example": "คุณเอ"
          },
          "contactPhone": {
            "type": "string",
            "example": "0812345678"
          },
          "confirmImmediately": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "serviceCode",
          "customerSiteId",
          "requestedDate",
          "requestedStart",
          "requestedEnd",
          "customerOrganizationId"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "example": 1
          },
          "serviceCode": {
            "type": "string",
            "example": "GREASE_TRAP"
          },
          "customerSiteId": {
            "type": "string",
            "format": "uuid"
          },
          "requestedDate": {
            "type": "string",
            "example": "2026-09-10"
          },
          "requestedStart": {
            "type": "string",
            "example": "10:00"
          },
          "requestedEnd": {
            "type": "string",
            "example": "12:30"
          },
          "estimatedVolume": {
            "type": "number",
            "example": 10
          },
          "volumeUnit": {
            "type": "string",
            "example": "ลบ.ม."
          },
          "customerNote": {
            "type": "string",
            "example": "เข้าทางประตู 2"
          },
          "internalNote": {
            "type": "string",
            "example": "โทรยืนยันแล้ว ลูกค้าขอเลื่อนเวลา"
          },
          "responsibleUserId": {
            "type": "string",
            "format": "uuid"
          },
          "changeReason": {
            "type": "string",
            "example": "ลูกค้าขอเลื่อนเวลาและเพิ่มปริมาณ"
          },
          "assignmentResolution": {
            "type": "string",
            "example": "UNASSIGN_IF_INVALID",
            "enum": [
              "UNASSIGN_IF_INVALID",
              "CANCEL_EDIT"
            ]
          },
          "override": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "version",
          "changeReason"
        ]
      },
      "BookingActionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "ลูกค้าขอยกเลิกทางโทรศัพท์"
          }
        },
        "required": [
          "reason"
        ]
      },
      "AssignBookingDto": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "driverUserId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "example": "รถคันเดิมเข้าซ่อม"
          }
        },
        "required": [
          "vehicleId"
        ]
      },
...
    }
  }
}