# packages/api-client/src/generated/model/createOpsBookingDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Foresee Reservation API
 * Booking and operations API for Foresee Corporation
 * OpenAPI spec version: 1.0
 */
import type { CreateOpsBookingDtoSource } from './createOpsBookingDtoSource';

export interface CreateOpsBookingDto {
  serviceCode: string;
  customerSiteId: string;
  requestedDate: string;
  requestedStart: string;
  requestedEnd: string;
  estimatedVolume?: number;
  volumeUnit?: string;
  customerNote?: string;
  customerOrganizationId: string;
  source?: CreateOpsBookingDtoSource;
  internalNote?: string;
  responsibleUserId?: string;
  contactName?: string;
  contactPhone?: string;
  confirmImmediately?: boolean;
}