# packages/api-client/src/generated/model/updateBookingDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Foresee Reservation API
 * Booking and operations API for Foresee Corporation
 * OpenAPI spec version: 1.0
 */
import type { UpdateBookingDtoAssignmentResolution } from './updateBookingDtoAssignmentResolution';

export interface UpdateBookingDto {
  version: number;
  serviceCode?: string;
  customerSiteId?: string;
  requestedDate?: string;
  requestedStart?: string;
  requestedEnd?: string;
  estimatedVolume?: number;
  volumeUnit?: string;
  customerNote?: string;
  internalNote?: string;
  responsibleUserId?: string;
  changeReason: string;
  assignmentResolution?: UpdateBookingDtoAssignmentResolution;
  override?: boolean;
}